MultiMeasures MultiDimesions -Olatunji's Coffee
This image shows Multidimensions (Product Group, Product Category, Product) and MultiMeasures (Sales, Quantity Sold, Customers) used to provide the user with multiple combinations of measures and dimension values to break them down with. A dropdown list for each is provided for the user to view which combination of measure and dimension they desire.
Additionally, on the final Dashboard, a parameter driven sheet selector allows the user to view the measures and dimensions as different chart types to keep them engaged and better informed on the information they are seeing. These different charts use the same space on the dashboard. 
Full Dashboard to follow.....