Weekly Sales - Index Value
This image shows a dual chart of Sales (orange circles) and a MultiMeasure (Sales, Quantity Sold and Customers). The MultiMeasure is a Percent Difference and on the Dashboard is set to compare every other value to the selected one. This was done with a parameter of a fixed value that is then selected as the 'Relative To' value in the Percent Difference Table Calculation.
Full Dashboard to folow.....